// src/lib/types/admin-features.d.ts
export type ActiveSession = {
  user_id: number;
  first_name: string;
  last_name: string;
  username: string;
  email: string;
  role_name: string;
  session_expires_at: string;
  last_login_at: string | null;
  last_login_ip: string | null;
};